+++ Elements/src/Geometry/CsgExtensions.cs
@@ -27,17 +27,27 @@ namespace Elements.Geometry
         /// </summary>
         internal static GraphicsBuffers Tessellate(this Csg.Solid csg)
         {
-            return Tesselate(new[] { csg });
+            return Tessellate(new[] { csg });
         }
 
         /// <summary>
         /// Triangulate a collection of CSGs and pack the triangulated data into
         /// buffers appropriate for use with gltf. 
         /// </summary>
-        internal static GraphicsBuffers Tesselate(this Csg.Solid[] csgs)
+        internal static GraphicsBuffers Tessellate(this Csg.Solid[] csgs)
         {
             var buffers = new GraphicsBuffers();
 
+            Tessellate(csgs, buffers);
+            return buffers;
+        }
+
+        /// <summary>
+        /// Triangulate a collection of CSGs and pack the triangulated data into
+        /// a supplied buffers object. 
+        /// </summary>
+        internal static void Tessellate(Csg.Solid[] csgs, IGraphicsBuffers buffers)
+        {
             ushort iCursor = 0;
 
             (Vector3 U, Vector3 V) basis;
@@ -58,16 +68,16 @@ namespace Elements.Geometry
                 {
                     var vertexIndices = new ushort[p.Vertices.Count];
 
-                    var a = p.Vertices[0].Pos.ToElementsVector();
-                    var b = p.Vertices[1].Pos.ToElementsVector();
-                    var c = p.Vertices[2].Pos.ToElementsVector();
-                    basis = ComputeBasisAndNormalForTriangle(a, b, c, out Vector3 normal);
-
                     // Anything with 3 vertices is a triangle. Manually 
-                    // tesselate triangles. For everything else, use 
+                    // tessellate triangles. For everything else, use 
                     // the tessellator.
-                    if (p.Vertices.Count > 2 && p.Vertices.Count <= 3)
+                    if (p.Vertices.Count == 3)
                     {
+                        var a = p.Vertices[0].Pos.ToElementsVector();
+                        var b = p.Vertices[1].Pos.ToElementsVector();
+                        var c = p.Vertices[2].Pos.ToElementsVector();
+                        basis = ComputeBasisAndNormalForTriangle(a, b, c, out Vector3 normal);
+
                         for (var i = 0; i < p.Vertices.Count; i++)
                         {
                             var v = p.Vertices[i];
@@ -107,6 +117,14 @@ namespace Elements.Geometry
                         {
                             continue;
                         }
+                        // We pick the first triangle from the tesselator,
+                        // instead of the first three vertices, which are not guaranteed to be
+                        // wound correctly.
+                        var a = tess.Vertices[tess.Elements[0]].ToElementsVector();
+                        var b = tess.Vertices[tess.Elements[1]].ToElementsVector();
+                        var c = tess.Vertices[tess.Elements[2]].ToElementsVector();
+
+                        basis = ComputeBasisAndNormalForTriangle(a, b, c, out Vector3 normal);
 
                         for (var i = 0; i < tess.Vertices.Length; i++)
                         {
@@ -138,7 +156,6 @@ namespace Elements.Geometry
                     }
                 }
             }
-            return buffers;
         }
 
         private static (Vector3 U, Vector3 V) ComputeBasisAndNormalForTriangle(Vector3 a, Vector3 b, Vector3 c, out Vector3 n)
@@ -296,7 +313,7 @@ namespace Elements.Geometry
 
             if (p.Vertices.Count == 3)
             {
-                // Don't tesselate unless we need to.
+                // Don't tessellate unless we need to.
 
                 var a = p.Vertices[0];
                 var b = p.Vertices[1];
@@ -313,7 +330,7 @@ namespace Elements.Geometry
             }
             else if (p.Vertices.Count == 4)
             {
-                // Don't tesselate unless we need to.
+                // Don't tessellate unless we need to.
 
                 var a = p.Vertices[0];
                 var b = p.Vertices[1];
@@ -387,6 +404,11 @@ namespace Elements.Geometry
         private static Csg.Vector3D ToCsgVector3(this ContourVertex v)
         {
             return new Csg.Vector3D(v.Position.X, v.Position.Y, v.Position.Z);
+        }
+
+        private static Vector3 ToElementsVector(this ContourVertex v)
+        {
+            return new Vector3(v.Position.X, v.Position.Y, v.Position.Z);
         }
 
         private static ContourVertex[] ToContourVertices(this List<Csg.Vertex> vertices)