+++ Elements/src/Geometry/Vector3.cs
@@ -129,10 +129,11 @@ namespace Elements.Geometry
         /// <summary>
         /// Return a new vector which is the unitized version of this vector.
         /// </summary>
-        public Vector3 Unitized()
+        /// <param name="tolerance">The amount of tolerance in the zero length comparison.</param>
+        public Vector3 Unitized(double tolerance = Vector3.EPSILON)
         {
             var length = Length();
-            if (length == 0)
+            if (length.ApproximatelyEquals(0, tolerance))
             {
                 return this;
             }